Move the regions data
We move the data into a `db/seeds` and away from `test/support/seeders/data` so that the seeding and the seed data will live together.

We also make some changes to the contents. Though possible to export values directly we think it is clearer to first declare them then export them.

We also make changes to the seed data itself. We shouldn't de hard coding the created and updated at times as these will be set when the records are created. In our tests we sometime do need to set these. But only ever when dealing with transactional data. There is no scenario where the created or updated at values of a reference record matters.