Replace seeder
The original seeder was using raw SQL to perform the update/insert. However, there is no reason why we cannot use Objection.js as we do in the majority of our database interactions.

We've also changed what kind of queries get run. With no constraint on the table, we'll always need to run 2 queries. The original ran both the update and insert, each with clauses that would mean only one would actually make any changes.

On reflection, we think it is clearer and safer to only run the 'mutating' query needed. So, instead our first query is to determine if a region already exists. If it does, we perform an update. If it doesn't we perform the insert.

We've also extended the seeder to cater for the fact `water.regions` has an `is_test` column we don't expose in `public.regions`. Our existing tear down and acceptance tests depend on the 'Test Region' having this flag set. So, we cater for it in our new seeder.

The final change is the file name. Knex runs the seed files in the order found. Because of this it is considered good practise to name your seed files in such a way that you control the order they are run in. Hence the pattern of prefixing seed file names with numbers.