Update tear down service (#1232)
Following on from updates in the acceptance tests repo, all the data being used is now from a fixtures file in the repo (instead of using the legacy code). We need to update the tear-down services idm schema to include a new like clause when deleting data. This update means all the relevant test data gets deleted every time the acceptance tests are run.