Update acceptance test tear down
https://github.com/DEFRA/water-abstraction-acceptance-tests/pull/129

During the removal of region data in our acceptance test fixture file, it was noted that the new region seeder was breaking our tear down service. The tear down service relies on knowing if the region associated with the data is a test region or not through the 'isTest' property. The seeded test region has this set to false, when it needs to be true. This also means updating this data then deletes the test region itself at the end of ever tear down, so this PR is to update the region seeder and the tear down service.