Fix external ID in ret. req. purpose helper
Spotted when we update it to reference the new location for purpose data. We were still generating an external ID using the helper generate code methods but then selecting values from the source reference data.

An external ID is built from the codes of the purposes the return requirement purpose is linked to. So, it _should_ be in sync with the purposes actually linked.

We think there are other places where selecting a random entry from the reference data would also be needed. So, we also refactor some of the logic to a new helper method.