Update acceptance test tear down service (#1234)
DEFRA/water-abstraction-acceptance-tests#129

During the removal of our region data in the acceptance tests, it was noted that the new region seeder was being deleted in the tear-down service. The tear-down service relies on knowing if the region associated with the data is a test region or not through the 'isTest' property. After deleting the associated data with the test region it then deletes the region itself. As our region is now seeded and not being loaded fo every test, this last delete needs to be removed.
This PR is removing that.