Switch to using stubs to remove region dependence
We're about to update all references to `Region.add()` to be `Region.select` so that we are using the pre-existing reference data rather than adding duplicate records solely for a test.

This one test currently relies heavily on injecting data into the DB. However, it is calling 2 other services that handle fetching the data which have their own tests.

So, unlike the changes we're about to make, we can update this test to be 100% stub dependent.