Update and fix LoadService post RegionHelper change
Removing the `.add()` from RegionHelper broke the data LoadService used by our acceptance tests.

We already know that we no longer seed test regions through this mechanism from the acceptance tests. So, we can confidently remove it from the Loader. But it also means we have to use a different entity type in the unit tests (we switch to Licences) to get them passing again.