Create SROC two-part tariff flagging endpoint
https://eaflood.atlassian.net/browse/WATER-4622

This is apart of the work for two-part tariff supplementary billing. A licence can be added to a two-part tariff supplementary bill run in 5 ways. Editing a return, adding a historic charge version, removing a licence from the annual two-part tariff bill run, recalculate a bill and if a licence is lapsed or revoked. Out of these 5 ways, 4 of them are in the legacy code base. To keep the new code for the supplementary billing within our system code base, we have decided to create a new endpoint that the legacy code can hit to calculate if the changed licence needs to be added to a supplementary bill run. This PR is setting up the endpoint.