Create `notes` table model and helper
While creating the tests for the [Licence History page](https://github.com/DEFRA/water-abstraction-system/pull/1182),
we noticed we are missing the model and helper for the notes table. This is being referenced when extracting charge versions
for a licence and any relating notes created. As a result, while creating the tests for the `fetch-licence-history.service`
database errors occurred as a result of the unknown relation between the notes and charge versions table.

This PR will add the model and helper for the `notes` table in the water schema.