Removing licence from annual 2pt flags for supp billing (#1236)
https://eaflood.atlassian.net/browse/WATER-4590

When a licence is removed from an annual two-part tariff bill run, it must be flagged to be included in the supplementary two-part bill run. This PR sets the foundation for flagging a licence. Currently, on the licence table, we have a supplementary bill run flag used for pre-sroc supplementary bill runs. However, with a 2pt sroc one we need to be able to store the year that the licence needs to be included.

The solution is a new table called licenceSupplementaryYears. This table will contain details of the licenceId, billRunId (if its been picked up by a billRun), the financialYearEnd and if it is 2pt. We have the 2pt column to allow us in the future to change existing supplementary flagging for non-2pt bill runs. This PR is specifically looking at when a licence is removed from an annual bill run, and the details then get persisted in our new table.

It also looks at the scenario for when a licence is removed from an annual 2pt bill run but then the bill run is cancelled and created again.