Do not remove UserHelper add
We seed users so we have something to log in with when building a fresh environment. Our acceptance tests also depend on these users. But they are not 'true' reference data. It is expected that new users will be added and changes made to users within the service.

So, it is perfectly acceptable to create a new user as part of a unit test. We also need to retain the support in the loader for the acceptance tests to be able to create users as and when needed.

So, unlike the other helpers we keep the add() method for Users.