Reset Userhelper defaults() method
We've updated it in the oher helpers we've changed to use a selected entry. However, the method was built to support adding new records (we've left it for now just in case!)

With that in mind, in the case of UserHelper where we are trying to support both reference and transaction data, it is better to keep it generating dynamic new data, rather than selecting from an existing entry.