Update GroupRoles seeder to match UserGroups
It means dropping back to raw SQL rather than using Objection or Knex, and we don't get the benefit of using the tables `onConflict`.

But it keeps it consistent with UserGroups, and like UserGroups, it can properly handle seeding a populated DB and a fresh/test DB. There was still a possibility of picking the wrong group/role ID with the previous iteration.