Remove two-part tariff supplementary billing flag from the `Licences` view https://eaflood.atlassian.net/browse/WATER-4587
The PR https://github.com/DEFRA/water-abstraction-system/pull/1236 has changed the way the SROC two-part tariff supplementary billing flag is set. Previously there was a boolean flag in the `licences` table/view to indicate if the licence should be included in the next TPT supplementary bill run. This is no longer sufficient for our needs as we are now required to record the specific year that the TPT supplementary bill run is for.

So a new table `licence_supplementary_years` has been built that is now used to record if the licence should go into a supplementary bill run & for which year. As a result the original flag in the `licences` table is no longer required.

In this PR the `include_in_sroc_tpt_billing` column is going to be removed from the view. The next PR will be in the `water-abstraction-service` to remove the actual column from the table.