Attempt tp avoid collisions in IDM testing
We got intermittent failures in CI with the user testing, especially where we are having to create user-role records. This change reduces the number of times we are having to create them by moving all the setup of the user tests into a `before()` rather than `beforeEach()`. We also remove random record selection and ensure a spread of users, groups and roles across the tests.