Remove defaults() from reference-only helpers
Having ensured they stuck around for the UserHelper and UserGroupHelper because it made sense, it made us think again about their use in reference-only helpers.

They weren't being used and would be unlikely to. We've also found that in most cases we've been defining indexes in the tests. So, `DEFAULT_INDEX` as also not really needed.