Restructure 'seeding' to be Knex based (#1230)
https://eaflood.atlassian.net/browse/WATER-4575

We have been working on replicating how the [water-abstraction-import](https://github.com/DEFRA/water-abstraction-import) service imports a licence. See [Import Licence versions](https://github.com/DEFRA/water-abstraction-system/pull/1195) for example.

This is in readiness for moving to ReSP managing licences rather than NALD.

What it exposed was that our test data setup needed expanding. The import relies on 'real' reference data being present. So, it makes testing easier if this is the case in our test database when writing unit tests.

We started with [create wrapper for licence-version-purpose-condition table](https://github.com/DEFRA/water-abstraction-system/pull/1181) which added the ability to seed licence-purpose condition types. This was then expanded by [Add new test reference data seeders](https://github.com/DEFRA/water-abstraction-system/pull/1206) which added some more reference data, for example, purposes.

The issue we've had is that the existing changes are straddling two different mechanisms. We've included them as a seed file in our [Knex seeding capabilities](https://knexjs.org/guide/migrations.html#seed-files). This is reference data, so why not seed it?!

And we are seeding it whenever we perform a database clean. Our intent is to have real reference data when unit testing, and until we've removed calls to `DatabaseSupport.clean()` from all our tests, we need to do this each time. That has led to the logic and data itself living in the `test/` folder.

There is a conceptual problem with this: something we depend on 'for real' lives in our `test/` folder. Plus, the functionality is split into 2 places.

It also means our `test/support/seeders` have a mix of types. Original 'unit test seeders' that were intended to help move bulky test setup away from within the test itself, with 'knex seeders', things we'll also call when we run `npm run seed` against the non-test DB.

So, on reflection, we felt the need to carry out a little restructuring. _All_ the 'seeding' functionality is moved to `db/seeds` but with the ability to still call it from `clean()` until such time as we have updated all tests and can drop the command.

While we're doing this housekeeping, we also update our existing user seed to match the conventions that have been introduced while also re-implementing some existing ones, for example, preferring **Objection.js** to raw SQL queries.

The end result is that we can seed both the real and test databases with reference data, even when a database already has existing records, and this functionality will live in one place and be built on **Knex**.

But there's more!

Having made these changes to ensure both the real and test databases are correctly populated with reference data, it no longer made any sense to be adding new reference records within our tests.

So, we updated the relevant test helpers to provide a `select()` in place of the `add()` method. This means you can select the details of a reference record without having to first create or query for it.

We then went through the affected tests and 'test-seeders' and swapped calls to `Helper.add()` with `Helper.select()`. We also performed some housekeeping on some of the tests we touched.