Mark Licence for Supplementary 2PT bill run after creating a historic return version https://eaflood.atlassian.net/browse/WATER-4621
When a new return version is created we need to check to see if it has impacted any financial years that have already been included in a two-part tariff bill run.

If any financial years are impacted, then the licence needs to be flagged for a two-part tariff supplementary bill run for the years that a 2PT bill run exists.