Fix all remaining lint issues
As part of ongoing automation for our code practices and preferences we introduced eslint instead of the DEFRA standard linter standard.js.

Our eslint setup uses standard.js as the base linter but we have our own rules on top.

We have previously introduced a gradual adoption of eslint by using the git diff when running eslint. This appears to have some limitation with some rules not getting picked up.

At this point it makes sense to fix the remaining lint issues and remove the git diff and lint the whole codebase. When we first moved over to eslint there were 1000+ issues now we only have 137 issues (with the majority being able to be fixed with --fix).

This change fixes all outstanding linting issues and remove the previous lint with git diff lint to the basic eslint run.