Fix order of return versions in view licence
https://eaflood.atlassian.net/browse/WATER-4407

> Part of the work to replace NALD for handling return requirements

Spotted in testing is that if a licence has 2 return versions with the same start date (one will be 'approved' and the other will display as 'replaced'), it is possible that they will be ordered incorrectly. In this scenario, the 'approved' should be first in the table and then the 'replaced'.

We had exactly the [same issue with charge versions](https://github.com/DEFRA/water-abstraction-system/pull/1162). The solution was to order them first by start date, then by version number. So, we apply the same sorting to when we fetch the return versions.