Revert changes to charge and licence versions
We'll also be mapping the mod log reason to WRLS reasons at some point. Return versions is simple; it has a reason field. Charge versions links to a reference record so will need a bit extra. Licence has neither so will need some additional changes made to the table.

In light of that, we've decided to reduce the scope of the change to just the return version. That way we'll know better whether further changes are needed when we come to amend those migrations.