Add mod_log field to return versions view (#1252)
https://eaflood.atlassian.net/browse/WATER-4565

> Part of the work to display a licence's history to users (mod log)

NALD has a concept called 'mod log'. Essentially, when someone creates a new licence, charge, or return version, they are required to provide a reason and can also add a note.

Rather than this being stored against each record, the 'mod log' is stored in a single place. Users can then view a licence's 'mod logs' to get a sense of the history of the licence.

When charging switched from NALD to WRLS, we (the previous team) built the ability to record a reason and note against a new charge version but didn't import any historic data. With return requirements also soon to switch from NALD to WRLS, there is a concern that this view of changes made to a licence will be lost.

Having reviewed how the information is held in NALD, we've confirmed that we can extract this information and import it as part of the licence import. The first step is ensuring we have somewhere to put it.

This change adds a migration to alter the return versions view to include the new `mod_log` field we've added to the underlying table in the `water` schema.