Remove package.json from the JSDoc config
Having tested it what it does is generate output specific to the project name and version inside the `docs/` folder. This would useful if you wanted to generate a version of your documentation for each release of your project.

That is not something we are interested in. We are just trying to validate we have our JSDoc comments correct within the code. Removing it means the documentation is more easily accessible at `docs/index.html`.