Add JSDoc as dependency to validate documentation (#1253)
https://github.com/DEFRA/water-abstraction-team/issues/125

Using [JSDoc comments](https://jsdoc.app/) to document our code has become our convention.

We add `@module` declarations by default, detail our public methods, including params and returns, and have even added markdown sections in some places.

Recently, we went on a deep dive of **JSDoc** to generate documentation based on what we've added so far. The good news is that it confirmed that we've got a lot of stuff right! But there are some things we need to correct, plus some errors that make comments invalid.

This change adds **JSDoc** as a dev dependency to the project. We can then use it to generate documentation on request. We then add it to our CI, where it will error when it encounters documentation it cannot parse, which serves as a means to validate that what we've done is 'error-free'.

To ensure CI doesn't fail after this change, we correct our current errors.

From now on, we'll catch any parsing errors. Hopefully, generating the documentation will also allow us to view how what we've added is presented. Having carried out this work, we know there are some issues to be fixed, but those are for another PR!