Trigger 2PT Supplementary Bill Run https://eaflood.atlassian.net/browse/WATER-4200
As we are treating 2PT supplementary as a separate bill run we will need a new journey to trigger it.

In this PR a link for 2PT Supplementary will be added to the list of bill runs that can be triggered.  Selecting this will take you to a screen to select the region.