Update view return version to include mod log info
https://eaflood.atlassian.net/browse/WATER-4634

With the information around modLog in WRLS, we intend to display it in the
new [licence history page](https://eaflood.atlassian.net/browse/WATER-4566). This
information being showed on the history page is the same as what is being displayed
on the return version page. So we should update it to use mod log info to make
the page more accurate.

This PR will incorporate the modLog information into the return version page.