Rename mod_log field in version views
https://eaflood.atlassian.net/browse/WATER-4565

> Part of the work to display a licence's history to users (mod log)

In [Add mod_log field to return versions view](https://github.com/DEFRA/water-abstraction-system/pull/1252), we added a migration to add the new `mod_log` field to the charge, licence, and return version views. This field was intended to hold details from the linked NALD mod log record, such as who, when, and why the version was created.

However, we've since learned that a version in NALD can have multiple mod log records. Therefore, we want to change the name to `mod_logs` to make that clear.

Because the previous migrations have not yet been released, we'll update them rather than add new ones.

We also overlooked telling Objection.js that the field is JSON in the models. So, we do that as part of this change as well.