Remove return versions view mod log change (#1263)
https://eaflood.atlassian.net/browse/WATER-4565

> Part of the work to display a licence's history to users (mod log)

In [Add mod_log field to return versions view](https://github.com/DEFRA/water-abstraction-system/pull/1252), we added a migration to add the new `mod_log` field to the return version view. This field was intended to hold details from the linked NALD mod log record, such as who, when, and why the version was created.

However, we've since learned that a version in NALD can have multiple mod log records. So, let's rename the field `mod_logs` and change its default to an array. 😜

**It's a trap!**

We have made the same mistake as the previous team. When we believed that a version (charge, licence, or return) had a one-to-one relationship with NALD mod log records, a JSONB field in each was a suitable solution to capturing just the information we needed. After all, it's only the historical records we have to worry about.

But when we learned that a version record could be linked to multiple mod logs, we should have taken a different approach. Instead, we ploughed on with our JSON column [altering its name and default](https://github.com/DEFRA/water-abstraction-service/pull/2612) to match. 🤦😬

Once you've started down the JSONB road, it is easy to get trapped!

Thankfully, we've come to our senses before this saw the light of day in production. The mod log data we are importing needs to go into its own table, and we will link our version records accordingly. A 'standard' relationship database solution.

We've made a change to [water-abstraction-service to delete the column migrations and add a new table one instead](https://github.com/DEFRA/water-abstraction-service/pull/2613). This is because of the interaction between our views and the tables: you cannot drop a column referenced by a view.

That means we need to do the same thing here: delete the migration we added and pretend it never happened!