Add history attributes to ReturnVersionModel
https://eaflood.atlassian.net/browse/WATER-4560

> Part of the work to display NALD mod log records as history in WRLS

So, we are now importing mod logs from NALD and have added a model linked to charge, licence and return version models to allow us to retrieve them from the database.

We are in the process of building [a new licence history page](https://github.com/DEFRA/water-abstraction-system/pull/1182) and [Updating view return version to include mod log info](https://github.com/DEFRA/water-abstraction-system/pull/1261). These are 2 places that will need to display the same 'history' attributes, for example, when the record was created and who by, in two different places.

Because of this we believe there is value in adding a modifier and instance method to the `ReturnVersionModel` that will handle fetching the mod logs and determining this information.

This change updates the `ReturnVersionModel` with these properties. This will also form a pattern we'll apply to `LicenceVersionModel` and `ChargeVersionModel` in later changes.