Add history attributes to ReturnVersionModel (#1267)
https://eaflood.atlassian.net/browse/WATER-4560

> Part of the work to display NALD mod log records as history in WRLS

So, we are now importing mod logs from NALD and have added a model linked to charge, licence and return version models to allow us to retrieve them from the database.

We are in the process of building [a new licence history page](https://github.com/DEFRA/water-abstraction-system/pull/1182) and [Updating view return version to include mod log info](https://github.com/DEFRA/water-abstraction-system/pull/1261). These two places will need to display the same 'history' attributes, for example, when the record was created and who created it.

Because of this, we believe adding a modifier and instance methods to the `ReturnVersionModel` that will handle fetching the mod logs and determining this information is valuable.

This change updates the `ReturnVersionModel` with these properties. In later changes, this will also form a pattern we'll apply to `LicenceVersionModel` and `ChargeVersionModel`.